descrease-batch-size

+++ DatasetAdv/translate.py
@@ -72,8 +72,8 @@ def main():
             with accelerator.accumulate(model):
                 # Convert images to latent space
                 inputs = batch['text']
-                inputs['input_ids'] = inputs['input_ids'].squeeze(1).to(accelerator.device).to(weight_dtype)
-                inputs['attention_mask'] = inputs['attention_mask'].squeeze(1).to(accelerator.device).to(weight_dtype)
+                inputs['input_ids'] = inputs['input_ids'].squeeze(1).to(accelerator.device).long()
+                inputs['attention_mask'] = inputs['attention_mask'].squeeze(1).to(accelerator.device).long()
                 output_ids = accelerator.unwrap_model(model).generate(
                     **inputs,
                     decoder_start_token_id=test_dataset.tokenizer.lang_code_to_id["vi_VN"],
